fix: restore and reorder time test
